refactor: move AlertListScreenTest.kt to correct directory, move AlertList.kt to correct directory
